moved components hierarchy to App.js, react folder to react, cleaned up old icons
moved components hierarchy to App.js, react folder to react, cleaned up old icons, moved Text Component into Components folder
