+++ ui/react/package.json
@@ -8,7 +8,8 @@
     "@testing-library/user-event": "^7.2.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-scripts": "3.3.0"
+    "react-scripts": "3.3.0",
+    "react-switch": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
