+++ ui/react/src/components/TextComponent.js
