+++ ui/react/src/index.js
@@ -3,11 +3,12 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import TextComponent from './components/TextComponent';
 import ToggleSwitch from './components/toggleSwitch';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-ReactDOM.render(<ToggleSwitch />, document.getElementById('root'));
-
+//ReactDOM.render(<ToggleSwitch />, document.getElementById('root'));
+//ReactDOM.render(< TextComponent/>, document.getElementById('root'));
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
